Added bettervannillabuilding
BetterVanillaBuilding is a texturepack that uses optifine connected textures in a way like never seen before. Giving its users access to a big range of new building blocks, as well as new block variations in biomes to explore, new mobs to look at and new armors and tools to wield!